Update README for moved sample binaries
