+++ tests/sample/bin_impl.rs
