+++ tests/sample_deferred/bin_impl.rs
