+++ tests/sample_shadows/bin_impl.rs
